requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.6
channels==3.0.3
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.5
django-admin-honeypot==1.1.0
django-countries==7.2.1
django-filter==2.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Pillow==8.3.1
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
zope.interface==5.4.0
